Thomas McCarthy Project
This is my Computer Science project. I used html coding to make it. It is about me and my family. It has picutres in it which i coded in. It has serval lists in it, which i made. I changed to font, text color, and background color. I made a table which is seen on the home screne